+++ src/lib/components/admin/Settings/Connections/OllamaConnection.svelte
@@ -58,7 +58,7 @@
 
 <ManageOllamaModal bind:show={showManageModal} urlIdx={idx} />
 
-<div class="flex gap-1.5">
+<div class="ow-settings-row flex w-full gap-1.5 items-center">
 	<Tooltip
 		className="w-full relative"
 		content={$i18n.t(`WebUI will make requests to "{{url}}/api/chat"`, {